+++ database/migrations/2021_11_11_110838_remove_visit_count_in_branch_visitor_histories_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class RemoveVisitCountInBranchVisitorHistoriesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('branch_visitor_histories', function (Blueprint $table) {
+
+            $table->dropColumn('visit_count');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('branch_visitor_histories', function (Blueprint $table) {
+
+            $table->unsignedInteger('visit_count')->after('visitor_id');
+        });
+    }
+}